// src/lib/api/generated/models/adminOrdersControllerFindAllOrderStatus.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * Beauty Pack Recommendation API
 * Backend API for beauty-pack personalization, recommendations, COD orders, and administration
 * OpenAPI spec version: 1.0.0
 */

export type AdminOrdersControllerFindAllOrderStatus = typeof AdminOrdersControllerFindAllOrderStatus[keyof typeof AdminOrdersControllerFindAllOrderStatus];


export const AdminOrdersControllerFindAllOrderStatus = {
  PENDING_CONFIRMATION: 'PENDING_CONFIRMATION',
  CONFIRMED: 'CONFIRMED',
  PREPARING: 'PREPARING',
  SHIPPED: 'SHIPPED',
  DELIVERED: 'DELIVERED',
  CANCELED: 'CANCELED',
  RETURNED: 'RETURNED',
} as const;