// src/lib/api/generated/models/adminOrdersControllerFindAllSourceChannel.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * Beauty Pack Recommendation API
 * Backend API for beauty-pack personalization, recommendations, COD orders, and administration
 * OpenAPI spec version: 1.0.0
 */

export type AdminOrdersControllerFindAllSourceChannel = typeof AdminOrdersControllerFindAllSourceChannel[keyof typeof AdminOrdersControllerFindAllSourceChannel];


export const AdminOrdersControllerFindAllSourceChannel = {
  INSTAGRAM: 'INSTAGRAM',
  WHATSAPP: 'WHATSAPP',
  TIKTOK: 'TIKTOK',
  FACEBOOK: 'FACEBOOK',
  DIRECT: 'DIRECT',
  OTHER: 'OTHER',
} as const;